32bit環境でビルドする際に発生する警告の修正

+++ src/lib/coil/vxworks/coil/SharedMemory.cpp
@@ -174,7 +174,7 @@ namespace coil
 	  {
 		  return -1;
 	  }
-	  memcpy(&m_shm[pos],&data[0],size);
+	  memcpy(&m_shm[pos],&data[0],static_cast<size_t>(size));
 
     
 	  return 0;
@@ -207,7 +207,7 @@ namespace coil
 	  {
 		  return -1;
 	  }
-	  memcpy(&data[0],&m_shm[pos],size);
+	  memcpy(&data[0],&m_shm[pos],static_cast<size_t>(size));
 
 	  return 0;
   }
